Rehosted on github. Update to 1:1.1
I assumed configshell was version 1.99, like rtslib and targetcli, but
unfortunately it is not. Since we have already produced RPMs for rhel
and rawhide with 1.99, we need to increment Epoch to 1 to go back to
version 1.1, as tagged in git.

Rehost onto github. This allows us to package based on a proper tarball,
and allows intermediate upstream releases by moi. (This is necessary because
RTS will only take MIT, not AGPLv3, contributions.)

Signed-off-by: Andy Grover <agrover@redhat.com>